utils/python/requirements.txt
requests==2.19.1
psycopg2==2.7.5
PyYAML==3.13